Enable -Wcompat=error in the testsuite
Summary:

Enabling -Werror=compat in the testsuite allows us to easily see the impact
that a new warning has on code. It also means that in the period between adding
the warning and making the actual breaking change, all new test cases that are
being added to the testsuite will be forwards-compatible. This is good because
it will make the actual breaking change contain less irrelevant testsuite
updates.

Things that -Wcompat warns about are things that are going to break in the
future, so we can be proactive and keep our testsuite forwards-compatible.

This patch consists of three main changes:

* Add `-Wstar-is-type` to `-Wcompat`, in accordance with https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0030-remove-star-kind.rst
* Add `TEST_HC_OPTS += -Werror=compat` to the testsuite configuration.
* Fix all broken test cases.

Test Plan: Validate

Reviewers:

Subscribers:

GHC Trac Issues: #15278